# Improve performance and reduce memory footprint during generation of complex trees