@luketimothyjones 2021-06-28
If you want to run your own tests, I used the decorator function below to run accurate benchmarks. The code is a modified version of the SO answer [here](https://stackoverflow.com/a/48923087).

```python
import time
import os
import functools

def track(func):
    
    @functools.wraps(func)
    def wrapper(*args, **kwargs):
        start = time.process_time()
        result = func(*args, **kwargs)
        elapsed_time = time.process_time() - start

        print("\n{}: exec time: {}\n".format(
            func.__name__,
            elapsed_time))
        return result

    return wrapper