@friggog 2021-06-29
Nice! Do we need to bump the version for a new release?