@luketimothyjones 2021-06-29
I mean, it's a substantial improvement but ultimately a minor change. If we use the major.minor.patch format for versioning I guess it would count as a minor increment.